Added functionality to CreateEvent
Create event now creates an event in the database. It takes hard coded
time and location data. It does not check the user's data to be correct.
And the userID is hardcoded to be my id just for the moment. Also it
doesn't close the view when you press save (so dont spamm the save
button!)